// src/shared/api/models/getApiDoctorsPortalDoctors200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiDoctorsPortalDoctors200DoctorsItem } from './getApiDoctorsPortalDoctors200DoctorsItem';

export type GetApiDoctorsPortalDoctors200 = {
    currentPage: number;
    totalPages: number;
    limit: number;
    total: number;
    doctors: GetApiDoctorsPortalDoctors200DoctorsItem[];
};